feat(object): introduce property descriptors and prototype
Refactor the internal object model to support property descriptors
(`DataDescriptor`, `AccessorDescriptor`) and a `prototype` map for
`JsObject` and `JsArray`. This enables more spec-compliant property
definition and inheritance.

`make_object` and `make_array` are moved to `jspp::Object` and updated
to use the new system. `js_value_to_string` is renamed to `to_string`.