fix(runtime): prevent new operator on non-constructors
Introduce an `is_constructor` flag for `JsFunction` to accurately determine if a function can be used with the `new` operator. Pass the function's name to `AnyValue::construct` to provide more descriptive `TypeError` messages when attempting to construct a non-constructor. This specifically fixes `new Symbol()` which now correctly throws a `TypeError`.